refactor(core): de-angularize ManifestReader
