+++ app/scripts/modules/core/src/manifest/ManifestReader.ts
