--- app/scripts/modules/core/src/manifest/manifestWriter.module.ts
@@ -1,5 +1,0 @@
-import { module } from 'angular';
-import { MANIFEST_WRITER } from 'core/manifest/manifestWriter.service';
-
-export const MANIFEST_MODULE = 'spinnaker.core.manifest';
-module(MANIFEST_MODULE, [MANIFEST_WRITER]);
