+++ app/scripts/modules/core/src/storage/index.ts
@@ -1,1 +1,1 @@
-export * from './storageAccount.service';
+export * from './StorageAccountReader';
