+++ app/scripts/modules/kubernetes/src/v2/manifest/rollout/undo.controller.ts
@@ -1,7 +1,7 @@
 import { copy, IController, module } from 'angular';
 import { IModalServiceInstance } from 'angular-ui-bootstrap';
 
-import { Application, MANIFEST_WRITER, ManifestWriter, TaskMonitor } from '@spinnaker/core';
+import { Application, ManifestWriter, TaskMonitor } from '@spinnaker/core';
 import { IManifestCoordinates } from '../IManifestCoordinates';
 
 interface IUndoRolloutCommand {
@@ -28,7 +28,6 @@ class KubernetesManifestUndoRolloutController implements IController {
     coordinates: IManifestCoordinates,
     public revisions: IRolloutRevision[],
     private $uibModalInstance: IModalServiceInstance,
-    private manifestWriter: ManifestWriter,
     private application: Application,
   ) {
     'ngInject';
@@ -61,14 +60,14 @@ class KubernetesManifestUndoRolloutController implements IController {
       const payload = copy(this.command) as any;
       payload.cloudProvider = 'kubernetes';
 
-      return this.manifestWriter.undoRolloutManifest(payload, this.application);
+      return ManifestWriter.undoRolloutManifest(payload, this.application);
     });
   }
 }
 
 export const KUBERNETES_MANIFEST_UNDO_ROLLOUT_CTRL = 'spinnaker.kubernetes.v2.manifest.undoRollout.controller';
 
-module(KUBERNETES_MANIFEST_UNDO_ROLLOUT_CTRL, [MANIFEST_WRITER]).controller(
+module(KUBERNETES_MANIFEST_UNDO_ROLLOUT_CTRL, []).controller(
   'kubernetesV2ManifestUndoRolloutCtrl',
   KubernetesManifestUndoRolloutController,
 );
