+++ app/scripts/modules/kubernetes/src/v2/manifest/scale/scale.controller.ts
@@ -1,7 +1,7 @@
 import { copy, IController, module } from 'angular';
 import { IModalServiceInstance } from 'angular-ui-bootstrap';
 
-import { Application, MANIFEST_WRITER, ManifestWriter, TaskMonitor } from '@spinnaker/core';
+import { Application, ManifestWriter, TaskMonitor } from '@spinnaker/core';
 import { IManifestCoordinates } from '../IManifestCoordinates';
 import { KUBERNETES_SCALE_MANIFEST_SETTINGS_FORM } from './scaleSettingsForm.component';
 
@@ -24,7 +24,6 @@ class KubernetesManifestScaleController implements IController {
     coordinates: IManifestCoordinates,
     currentReplicas: number,
     private $uibModalInstance: IModalServiceInstance,
-    private manifestWriter: ManifestWriter,
     private application: Application,
   ) {
     'ngInject';
@@ -57,14 +56,14 @@ class KubernetesManifestScaleController implements IController {
       const payload = copy(this.command) as any;
       payload.cloudProvider = 'kubernetes';
 
-      return this.manifestWriter.scaleManifest(payload, this.application);
+      return ManifestWriter.scaleManifest(payload, this.application);
     });
   }
 }
 
 export const KUBERNETES_MANIFEST_SCALE_CTRL = 'spinnaker.kubernetes.v2.manifest.scale.controller';
 
-module(KUBERNETES_MANIFEST_SCALE_CTRL, [MANIFEST_WRITER, KUBERNETES_SCALE_MANIFEST_SETTINGS_FORM]).controller(
+module(KUBERNETES_MANIFEST_SCALE_CTRL, [KUBERNETES_SCALE_MANIFEST_SETTINGS_FORM]).controller(
   'kubernetesV2ManifestScaleCtrl',
   KubernetesManifestScaleController,
 );
